console: use interrupts for rx instead of dma
For whatever reason, DMA transfers aren't working. Instead, only the last
few characters are received. This appears to be a fault in the DMA
configuration, because manually-filled transfers work just fine.

Switch to an interrupt approach. This has no issues with partial transfers
and otherwise seems to work as well as the DMA approach.

Signed-off-by: Sean Cross <sean@xobs.io>